requirements.txt
aiofiles==23.2.1
aiogram==3.13.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiomysql==0.2.0
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
frozenlist==1.4.1
greenlet==3.1.0
idna==3.10
loguru==0.7.2
magic-filter==1.0.12
multidict==6.1.0
mysql-connector-python==9.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyMySQL==1.1.1
python-dotenv==1.0.1
PyYAML==6.0.2
SQLAlchemy==2.0.34
typing_extensions==4.12.2
win32-setctime==1.1.0
yarl==1.11.1
